moved errors under a proper type - maybe not done with error handling
The reason for this is to fix ZLS frequently crashing when it tries to
autocomplete errors when typing `return error.<whatever the error>`.

The error types I've defined are named according to the file in which
the errors are returned. There's probably a better way to do this, but
for now this works.

I also turned logging off by default - this is only for debugging right
now.